+++ src/components/common/MenuBar.js
@@ -12,27 +12,19 @@ const MenuBar = () => {
   return (
     <BottomBlock>
       <Link to="/" style={{ textDecoration: 'none', color: 'black' }}>
-        <NavHomeBlock>
-          <p>홈</p>
-        </NavHomeBlock>
+        <NavHomeBlock />
       </Link>
 
       <Link to="/write" style={{ textDecoration: 'none', color: 'black' }}>
-        <NavWrtieBlock>
-          <p>글쓰기</p>
-        </NavWrtieBlock>
+        <NavWrtieBlock />
       </Link>
 
       <Link to="/gps" style={{ textDecoration: 'none', color: 'black' }}>
-        <NavNearBlock>
-          <p>내 근처</p>
-        </NavNearBlock>
+        <NavNearBlock />
       </Link>
 
       <Link to="/mydanggeun" style={{ textDecoration: 'none', color: 'black' }}>
-        <NavMyDaangnBlock>
-          <p>나의 당근</p>
-        </NavMyDaangnBlock>
+        <NavMyDaangnBlock />
       </Link>
     </BottomBlock>
   );
