+++ src/containers/EditContainer.js
@@ -1,0 +1,9 @@
+import { useSelector, useDispatch } from 'react-redux';
+import { modify } from '../modules/profile';
+import { NameEdit } from '../styles/ProfileEditStyle';
+
+const EditContainer = () => {
+  const dispatch = useDispatch();
+};
+
+export default EditContainer;
