+++ src/pages/MainList.js
@@ -1,11 +1,9 @@
 import React from 'react';
-import MenuBar from '../components/common/MenuBar';
 
 const MainList = () => {
   return (
     <>
       <h1>Item_List</h1>
-      <MenuBar />
     </>
   );
 };
