+++ src/pages/ProfilePage.js
@@ -1,43 +1,110 @@
+// 나의당근 -> 프로필 화면
+
 import React from 'react';
 import { Link } from 'react-router-dom';
-import MenuBar from '../components/common/MenuBar';
-import ContainerModify from '../containers/ContainerModify';
-import { Danggeun } from '../styles/LoginStyle';
 import {
-  AttentionBlock,
-  BackImage,
-  ProfileImage,
-  ProfileModify,
-  ProfileName,
-  SaleBlock,
   TopBlock,
-  TopText,
+  ProfileImage,
+  BackArrow,
+  Profile,
+  ShareBlock,
+  MenuBlock,
+  ProfileName,
+  MannerBlock,
+  MannerImage,
+  ListLine,
+  BadgeBlock,
+  BadgeArrow,
+  SaleBlock,
+  SaleArrow,
+  TownBlock,
+  TownArrow,
+  Manner,
+  MannerArrow,
+  MannerText,
+  Temperature,
+  TemperatureImage,
+  BlueHr,
+  Heart,
+  ReviewBlock,
+  ReviewArrow,
+  ReviewText,
+  HeartText,
+  ChatBlock,
+  ChatText,
 } from '../styles/ProfileStyle';
 
 const ProfilePage = () => {
   return (
     <>
-      <Link to="/">
-        <BackImage />
+      <Link to="/mydanggeun">
+        <BackArrow />
       </Link>
       <TopBlock />
-
-      <TopText>프로필</TopText>
-      <ProfileImage />
+      <Profile>프로필</Profile>
+      <ShareBlock />
       <Link to="/profilemodify">
-        <ProfileModify>수정</ProfileModify>
+        <MenuBlock />
       </Link>
 
-      <ProfileName />
+      <ProfileImage />
+      <ProfileName>성윤</ProfileName>
+      <MannerBlock>매너온도</MannerBlock>
+      <MannerImage />
 
-      <Link to="/sale">
-        <SaleBlock>판매상품</SaleBlock>
-      </Link>
-      <Link to="/attention">
-        <AttentionBlock>관심상품</AttentionBlock>
-      </Link>
-      <Danggeun />
-      <MenuBar />
+      <Temperature>
+        <p>첫 온도 36.5 C</p>
+        <hr />
+      </Temperature>
+      <BlueHr />
+      <TemperatureImage>
+        <p>36.5 C</p>
+      </TemperatureImage>
+      <Heart>
+        <p>재거래희망률 -%</p>
+      </Heart>
+      <HeartText>표시될 만큼 충분히 대화하지 않았어요</HeartText>
+
+      <ChatBlock>
+        <p>응답률 33%</p>
+      </ChatBlock>
+      <ChatText>보통 1시간 내에 응답</ChatText>
+
+      <ListLine>
+        <p>
+          안양6동 9회 인증
+          <br />
+          <br />
+          2020년 11월30일 가입 (최근 3일 이내 활동)
+        </p>
+      </ListLine>
+
+      <BadgeBlock>
+        <p>활동 배지2개</p>
+        <BadgeArrow />
+      </BadgeBlock>
+
+      <SaleBlock>
+        <p>판매상품1개</p>
+        <SaleArrow />
+      </SaleBlock>
+
+      <TownBlock>
+        <p>동네생활</p>
+        <TownArrow />
+      </TownBlock>
+
+      <Manner>
+        <p>받은 매너 평가</p>
+        <MannerArrow />
+      </Manner>
+      <MannerText>받은 매너 칭찬이 없습니다</MannerText>
+
+      <ReviewBlock>
+        <p>받은 거래 후기(0)</p>
+        <ReviewArrow />
+      </ReviewBlock>
+      <ReviewText>받은 후기가 없습니다.</ReviewText>
     </>
   );
 };
