+++ src/styles/Layout.js
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const Layout = styled.div`
+  padding: 0 18px;
+  height: 100%;
+`;
