+++ src/styles/Timer.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+export const TimerBlock = styled.div`
+  width: 100%;
+  height: 50px;
+
+  font-size: 18px;
+  text-align: center;
+`;
+
+export const TimerLineBlock = styled.div`
+  border: 1px solid #f0f0f0;
+`;
